Fix INT-BOM_driven_image_library_comparison_with_trivy

+++ cks/INT-Isolating_a_pod_with_gvisor/reset.sh
@@ -11,8 +11,8 @@ SSH_OPTIONS="-o StrictHostKeyChecking=no -o UserKnownHostsFile=/dev/null -o LogL
 echo "🗑️  Removing Gvisor locally"
 
 sudo apt-get remove --purge -y runsc > /dev/null
-sudo rm -f /etc/apt/sources.list.d/gvisor.list
-sudo rm -f /usr/share/keyrings/gvisor-archive-keyring.gpg
+[ -f /etc/apt/sources.list.d/gvisor.list ] && sudo rm -f /etc/apt/sources.list.d/gvisor.list
+[ -f /usr/share/keyrings/gvisor-archive-keyring.gpg ] && sudo rm -f /usr/share/keyrings/gvisor-archive-keyring.gpg
 sudo apt-get autoremove -y  > /dev/null
 sudo apt-get update > /dev/null
 [ -f /etc/containerd/config.toml.SAVE ] && sudo mv /etc/containerd/config.toml.SAVE /etc/containerd/config.toml
@@ -23,13 +23,13 @@ for node in $(kubectl get nodes --no-headers | grep -v "control-plane" | awk '{p
   echo "🗑️  Removing Gvisor remotely on $node..."
   ssh $SSH_OPTIONS vagrant@$node bash <<'EOF'
     set -e
-	sudo apt-get remove --purge -y runsc > /dev/null
-	sudo rm -f /etc/apt/sources.list.d/gvisor.list
-	sudo rm -f /usr/share/keyrings/gvisor-archive-keyring.gpg
-	sudo apt-get autoremove -y  > /dev/null
-	sudo apt-get update > /dev/null
-  [ -f /etc/containerd/config.toml.SAVE ] && sudo mv /etc/containerd/config.toml.SAVE /etc/containerd/config.toml
-  sudo systemctl restart containerd
+    sudo apt-get remove --purge -y runsc > /dev/null
+    [ -f /etc/apt/sources.list.d/gvisor.list ] && sudo rm -f /etc/apt/sources.list.d/gvisor.list
+    [ -f /usr/share/keyrings/gvisor-archive-keyring.gpg ] && sudo rm -f /usr/share/keyrings/gvisor-archive-keyring.gpg
+    sudo apt-get autoremove -y  > /dev/null
+    sudo apt-get update > /dev/null
+    [ -f /etc/containerd/config.toml.SAVE ] && sudo mv /etc/containerd/config.toml.SAVE /etc/containerd/config.toml
+    sudo systemctl restart containerd
 EOF
 done
